centralized larus file reader in new folder larus_data, which is then used as a lib from analysis and flight _player.  Implemented usage of Larus2Df in flightplayer so that also *.f37 files can be imported there
