+++ larus_data/README.md
@@ -1,0 +1,4 @@
+# Larus Date
+- Python Lib to read the different binary larus fileformats into a pandas
+dataframe.
+- Used from analysis and flight_player
