+++ larus_data/_internal/README.md
@@ -1,0 +1,2 @@
+# Sensor data analyzer - binaries
+The files here are compiled versions of the ../sensor_data_analyzer/ for different platforms
