+++ larus_data/larus_to_df.py
@@ -1,19 +1,13 @@
 #!/user/bin/env python3
-"""
-This describes the different and some historical Larus file formats.
-"""
+import numpy as np
+import pandas as pd
+import os
+import sys
+import subprocess
+from pathlib import Path
 
 # Format with datatypes
-#Helper for generation
 # dtypes  int32_t  = i4,  int8_t = i1, uint16_t = u2,  f4 = float, f8 = double
-"""
-data = '['
-for element in elements_f120:
-    data += '(\"{}\", \"f4\"), '.format(element)
-data += ']'
-print(data)
-exit()
-"""
 
 # Raw data sensor files
 data_f37 = [("acc x", "f4"), ("acc y", "f4"), ("acc z", "f4"), ("gyro x", "f4"), ("gyro y", "f4"), ("gyro z", "f4"),
@@ -229,6 +223,73 @@ data_f123 = [("acc x", "f4"), ("acc y", "f4"), ("acc z", "f4"), ("gyro x", "f4")
              ("speed_comp_2", "f4"), ("speed_comp_3", "f4")]
 
 
-
 raw_data_formats = [('.f37', data_f37), ('.f50', data_f50)]
 processed_data_formats = [('.f110', data_f110), ('.f114', data_f114), ('.f120', data_f120), ('.f123', data_f123)]
+
+# Check if files has larus format file ending
+def check_if_larus_file(file):
+    for file_format in processed_data_formats:
+        if file.endswith(file_format[0]):
+            return True
+
+    for file_format in raw_data_formats:
+        if file.endswith(file_format[0]):
+            return True
+    return False # This is not a known Larus File format
+
+# Class to load a binary larus logfile into a pandas dataframe.  Converts raw data into processed data.
+class Larus2Df:
+    df = None
+    file = None
+    datatype = None
+    dataformat = None
+    def __init__(self, file):
+        for file_format in processed_data_formats:
+            if file.endswith(file_format[0]):
+                self.datatype = 'PROCESSED_DATA'
+                self.dataformat = file_format[1]
+
+        for file_format in raw_data_formats:
+            if file.endswith(file_format[0]):
+                self.datatype = 'RAW_DATA'
+                self.dataformat = file_format[1]
+
+        if self.datatype == 'RAW_DATA':
+            # Process raw data into processed data
+            if file.endswith('.f37'):
+                dataformat = data_f37
+                eeprom_file_path = str(file).replace('.f37', '.EEPROM')
+                result_file = str(file) + '.f114'
+
+                if not os.path.isfile(file):
+                    raise Exception("There must be a {} file".format(eeprom_file_path))
+
+                if not os.path.isfile(eeprom_file_path):
+                    raise Exception("There must be a {} file".format(eeprom_file_path))
+
+                if "linux" in sys.platform:
+                    subprocess.call(["{}/_internal/data_analyzer_commit_6598331_linux".format(Path(__file__).parent.absolute()), file])
+                elif "win" in sys.platform:
+                    subprocess.call(["{}/_internal/data_analyzer_commit_6598331_windows.exe".format(Path(__file__).parent.absolute()), file])
+                else:
+                    raise Exception("Platform not supported: {}".format(sys.platform))
+
+                file = result_file
+                self.dataformat = data_f114
+                self.datatype = 'PROCESSED_DATA'
+
+            else:
+                raise Exception("Raw data format {} not implemented yet.".format(self.dataformat))
+
+        # Convert larus dataformat into a pandas dataframe
+        dt = np.dtype(self.dataformat)
+        data = np.fromfile(file, dtype=dt, sep="")
+        self.df = pd.DataFrame(data)
+
+    def get_df(self):
+        return self.df
+
+
+if __name__ == "__main__":
+    df = Larus2Df('240520_091630.f37')
+    print(df.get_df().columns)
